+++ drools-drl-extensions/src/main/java/org/drools/drl/extensions/ResourceConversionResult.java
@@ -13,7 +13,7 @@
  * limitations under the License.
 */
 
-package org.drools.compiler.compiler;
+package org.drools.drl.extensions;
 
 import org.kie.api.io.ResourceType;
